update staff function

+++ Models/Record.cs
@@ -11,11 +11,11 @@ public partial class Record
 
     public bool Result { get; set; }
 
-    public decimal Co { get; set; }
+    public double Co { get; set; }
 
-    public decimal Hc { get; set; }
+    public double Hc { get; set; }
 
-    public decimal Nox { get; set; }
+    public double Nox { get; set; }
 
     public string Note { get; set; } = null!;
